backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
crewai==0.41.1
crewai[tools]==0.41.1
python-dotenv>=1.0.0
pydantic>=2.7.0,<3.0.0
requests>=2.31.0
langchain>=0.2.0,<=0.3.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
ibm-watson-machine-learning>=1.0.335
ibm-generative-ai>=3.0.0
